Proyecto Portafolio/requirements.txt
Flask==2.0.2
selenium==4.0.0
requests
beautifulsoup4
imageio
pillow

